feat: 입력값에 따라 구입 가능, 불가능 아이템 구분
